Move confirmation dialog to utils and ruff format
